Refactor UI into separate package
