+++ src/main/java/duke/command/AddEventCommand.java
@@ -1,7 +1,7 @@
 package duke.command;
 
 import duke.TaskList;
-import duke.Ui;
+import duke.ui.Ui;
 import duke.task.Event;
 import duke.task.Task;
