+++ src/main/java/duke/command/Command.java
@@ -2,7 +2,7 @@ package duke.command;
 
 import duke.DukeException;
 import duke.TaskList;
-import duke.Ui;
+import duke.ui.Ui;
 
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui) throws DukeException;
