+++ src/main/java/duke/command/DeleteCommand.java
@@ -2,7 +2,7 @@ package duke.command;
 
 import duke.DukeException;
 import duke.TaskList;
-import duke.Ui;
+import duke.ui.Ui;
 import duke.task.Task;
 
 public class DeleteCommand extends Command {
