+++ src/main/java/duke/command/FindCommand.java
@@ -2,7 +2,7 @@ package duke.command;
 
 import duke.DukeException;
 import duke.TaskList;
-import duke.Ui;
+import duke.ui.Ui;
 import duke.task.Task;
 
 import java.util.List;
