+++ src/main/java/duke/ui/Ui.java
@@ -1,5 +1,6 @@
-package duke;
+package duke.ui;
 
+import duke.DukeException;
 import duke.task.Task;
 
 import java.util.List;
